+++ app/js/webpack.config.js
@@ -1,0 +1,16 @@
+module.exports = {
+    entry: "index.js",
+    output: {
+        path: __dirname,
+        filename: "bundle.js"
+    },
+    module: {
+      loaders: [
+        {
+          test: /\.jsx?$/,
+          exclude: /(node_modules|bower_components)/,
+          loader: 'babel' // 'babel-loader' is also a legal name to reference
+        }
+      ]
+    }
+};